Added home page (#1195)
