+++ lib/common-lib/src/components/layout/HeaderTags/BoxBlue.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Box } from 'native-base'
 
-const BoxBlue = ({ children, ...props }) => {
+const BoxBlue = ({ children, width,height, ...props }) => {
   const styles = {
     box: {
       background:
@@ -11,8 +11,8 @@ const BoxBlue = ({ children, ...props }) => {
   return (
     <Box
       {...props}
-      width='200px'
-      height='143px'
+      width={width ? width : '200px'}
+      height={height ? height : '143px'}
       shadow='RedBlackShadow'
       // borderWidth='2px
       rounded='10px'
